refactor compute_topics and compute_vectors extract common functions
