+++ kg/src/pipe/compute_vectors.py
@@ -1,10 +1,10 @@
-import faiss
 from typing import Union
 
+import faiss
 import numpy as np
-import pandas as pd
 import torch
 import transformers
+from pipe import utils
 from pipe.data_checks.vectors import check_vector_size
 from transformers import AutoModel, AutoTokenizer
 from transformers.tokenization_utils_base import BatchEncoding
@@ -17,62 +17,6 @@ DATASET_PATH = ASSET_PATH + "dataset.pkl.gz"
 
 # Constants
 BATCH_SIZE = 1
-
-
-def create_inputs(title: str, abstract: str) -> str:
-    """
-    This function creates inputs
-    Args:
-        title: paper title
-        abstract: paper abstract
-    Return:
-        Clean and concatenated title + abstract with format:
-        "title. abstract"
-    """
-    title_abstract_str = f"{title}. {abstract}"
-    return title_abstract_str.replace("\n", " ").replace("**", " ")
-
-
-def get_inputs(input_path: str) -> list[str]:
-    """
-    Gets combination of title and abstracts for all the papers of the dataset.
-    Args:
-        input_path: dataset route
-    Return:
-        returns a list of strings with length (number_of_inputs)
-    """
-    df = pd.read_pickle(input_path, compression="gzip")[['title', 'abstract']]
-    return np.vectorize(create_inputs)(df['title'], df['abstract']).tolist()
-
-
-def convert_to_batches(input_items: Union[np.ndarray, list, range],
-                       batch_size: int) -> list[list[str]]:
-    """
-    Converts an iterable of items into a list of list
-    of items of the same size.
-
-    Args:
-        input_items: list of input items can be list of inputs range of indexes.
-        batch_size: batch size
-
-    Returns:
-        output: (list of list of list) containing input items in groups of batch_size
-
-
-    Example:
-        > convert_to_batches(list(range(0, 100)), batch_size=32)
-    """
-    n_batches = int(np.ceil(len(input_items) / batch_size))
-
-    output = [None] * n_batches
-
-    for batch_no in range(n_batches):
-        start = batch_no * batch_size
-        end = np.min([batch_no * batch_size + batch_size, len(input_items)])
-
-        output[batch_no] = input_items[start:end]
-
-    return output
 
 
 def tokenize_batch(samples: list[str],
@@ -159,8 +103,8 @@ def vectorize(inputs: list[str], batch_size: int,
         Returns an array of shape (number of inputs, vector_size)
     """
     # Convert inputs to batches
-    batches_of_docs = convert_to_batches(inputs, batch_size=batch_size)
-    batch_idx = convert_to_batches(range(len(inputs)), batch_size=batch_size)
+    batches_of_docs = utils.convert_to_batches(inputs, batch_size=batch_size)
+    batch_idx = utils.convert_to_batches(range(len(inputs)), batch_size=batch_size)
 
     tokens_batch_generator = (
         tokenize_batch(batch, tokenizer=tokenizer, tokenizer_args=tokenizer_args, device=device)
@@ -208,7 +152,7 @@ if __name__ == "__main__":
 
     # Vector computation
     print(f"Getting inputs from path {DATASET_PATH}")
-    dataset_inputs = get_inputs(DATASET_PATH)
+    dataset_inputs = utils.get_inputs(DATASET_PATH)
     print("Computing vectors")
     vectors = vectorize(inputs=dataset_inputs, batch_size=BATCH_SIZE,
                         model=TRANSFORMER, tokenizer=TOKENIZER,
